DAOS-9946 test: update ftest/ior/crash.py
Test-tag: ior_crash
Skip-unit-tests: true
Skip-fault-injection-test: true

- Rename ior/crash_ior to ior/crash
- Use 1 server, dual engine
- Use 3 clients
- Increase client processes per node
- Kill IOR in the middle of Read, not beginning
- General cleanup
