+++ src/tests/ftest/ior/crash.yaml
@@ -1,0 +1,56 @@
+hosts:
+  test_servers:
+    - server-A
+  test_clients:
+    - client-A
+    - client-B
+    - client-C
+timeout: 800
+server_config:
+  engines_per_host: 2
+  name: daos_server
+  servers:
+    0:
+      pinned_numa_node: 0
+      nr_xs_helpers: 1
+      fabric_iface: ib0
+      fabric_iface_port: 31317
+      log_file: daos_server0.log
+      log_mask: ERR
+      bdev_class: nvme
+      bdev_list: ["aaaa:aa:aa.a"]
+      scm_class: dcpm
+      scm_list: ["/dev/pmem0"]
+      scm_mount: /mnt/daos0
+    1:
+      pinned_numa_node: 1
+      nr_xs_helpers: 1
+      fabric_iface: ib1
+      fabric_iface_port: 31417
+      log_file: daos_server1.log
+      log_mask: ERR
+      bdev_class: nvme
+      bdev_list: ["bbbb:bb:bb.b"]
+      scm_class: dcpm
+      scm_list: ["/dev/pmem1"]
+      scm_mount: /mnt/daos1
+pool:
+  mode: 146
+  name: daos_server
+  size: 90%
+  control_method: dmg
+container:
+  type: POSIX
+  control_method: daos
+ior:
+  api: "DFS"
+  client_processes:
+    ppn: 32
+  dfs_destroy: False
+  flags: "-v -w -r -R"
+  sw_deadline: 120
+  test_file: /testFile
+  transfer_size: '4K'
+  block_size: '100G'
+  dfs_oclass: "SX"
+  subprocess: True
