--- src/tests/ftest/ior/crash_ior.yaml
@@ -1,44 +1,0 @@
-hosts:
-  test_servers:
-    - server-A
-    - server-B
-  test_clients:
-    - client-C
-timeout: 900
-server_config:
-  name: daos_server
-  servers:
-    log_mask: ERR
-    bdev_class: nvme
-    bdev_list: ["0000:81:00.0"]
-    scm_class: dcpm
-    scm_list: ["/dev/pmem0"]
-pool:
-  mode: 146
-  name: daos_server
-  scm_size: 6000000000
-  nvme_size: 100000000000
-  control_method: dmg
-container:
-  type: POSIX
-  control_method: daos
-ior:
-  api: "DFS"
-  client_processes:
-    np: 2
-  dfs_destroy: False
-  iorflags:
-    sequential:
-      flags: "-v -D 100 -w -r"
-  test_file: /testFile
-  repetitions: 1
-  transfersize_blocksize:
-    4K:
-      transfer_size: '4K'
-      block_size: '100G'
-  objectclass:
-    SX:
-      dfs_oclass: "SX"
-  subprocess: True
-dfuse:
-  mount_dir: "/tmp/daos_dfuse/"
